requirements.txt
djangorestframework==3.11.2
Django==3.1.8
Pygments==2.7.4
Markdown==3.2.1
coreapi==2.3.3
dj-database-url==0.5.0
gunicorn==20.0.4
whitenoise==5.0.1
django-cors-headers==3.9